Move FabricTable class out of transport layer
