+++ src/credentials/BUILD.gn
@@ -32,6 +32,8 @@ static_library("credentials") {
     "DeviceAttestationVendorReserved.h",
     "DeviceAttestationVerifier.cpp",
     "DeviceAttestationVerifier.h",
+    "FabricTable.cpp",
+    "FabricTable.h",
     "GenerateChipX509Cert.cpp",
     "GroupDataProvider.h",
     "examples/DeviceAttestationCredsExample.cpp",
