+++ src/credentials/tests/BUILD.gn
@@ -41,6 +41,7 @@ chip_test_suite("tests") {
     "TestChipCert.cpp",
     "TestDeviceAttestationConstruction.cpp",
     "TestDeviceAttestationCredentials.cpp",
+    "TestFabricTable.cpp",
     "TestGroupDataProvider.cpp",
   ]
