+++ src/transport/tests/BUILD.gn
@@ -23,7 +23,6 @@ chip_test_suite("tests") {
   output_name = "libTransportLayerTests"
 
   test_sources = [
-    "TestFabricTable.cpp",
     "TestPeerConnections.cpp",
     "TestSecureSession.cpp",
     "TestSessionHandle.cpp",
